Added support for command-line properties to Gradle plugin. Updated README.

+++ cloudfoundry-gradle-plugin/src/main/groovy/org/cloudfoundry/gradle/tasks/ScaleCloudFoundryTask.groovy
@@ -32,7 +32,7 @@ class ScaleCloudFoundryTask extends AbstractEnvCloudFoundryTask {
             log "Setting number of instances of application ${application} to ${instances}"
 
             withApplication {
-                client.updateApplicationInstances(application, instances)
+                client.updateApplicationInstances(application, instances as Integer)
             }
         }
     }
